pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pycroservice"
version = "0.0.7"
authors = [
  { name="inaimathi", email="leo.zovic@gmail.com" },
]
description = "A minimal Python microservice framework for Mycroft Technologies"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "Flask~=3.0",
    "Flask-Cors~=4.0",
    "pyjwt[crypto]~=2.8",
    "scoutsuite~=5.5",
    "gunicorn~=22.0"
]

[project.urls]
Homepage = "https://github.com/Mycroft-io/pycroservice"
Issues = "https://github.com/Mycroft-io/pycroservice/issues"